set up speed from point to point, animating properly along polyline

+++ Bento/Model/Step.h
@@ -14,8 +14,26 @@
 @property (nonatomic) CLLocationCoordinate2D startLocation;
 @property (nonatomic) CLLocationCoordinate2D endLocation;
 @property (nonatomic) NSMutableArray *pathCoordinates;
-@property (nonatomic) int duration;
+@property (nonatomic) NSInteger duration;
 
 - (id)initWithDictionary:(NSDictionary *)dictionary;
 
 @end
+
+// int count = 0
+
+// currentStep = steps[count]
+
+// currentStep.pathCoordinates
+
+// loop through currentStep.pathCoordinates with 1 second delay in between, animating from point to point while keeping count of the duration.
+
+// pathCoordinatesCount = 0
+
+// timer should repeatedly call a countPathCoordinates with 1 second delay
+
+// if pathCoordinatesCount <= currentStep.duration -> update driver with currentStep.pathCoordinates[pathCoordinatesCount]
+
+// 3 seconds to loop through 2 coordinates
+
+// 53 / 44
